+++ monads-tf/Control/Monad/RWS/Class.hs
@@ -31,6 +31,7 @@ import Control.Monad.State.Class
 import Control.Monad.Writer.Class
 
 import Control.Monad.Trans.Error(Error, ErrorT)
+import Control.Monad.Trans.Except(ExceptT)
 import Control.Monad.Trans.Maybe(MaybeT)
 import Control.Monad.Trans.Identity(IdentityT)
 import Control.Monad.Trans.RWS.Lazy as Lazy (RWST)
@@ -47,6 +48,8 @@ instance (Monoid w, Monad m) => MonadRWS (Strict.RWST r w s m)
 -- Instances for other mtl transformers
 
 instance (Error e, MonadRWS m) => MonadRWS (ErrorT e m)
+
+instance (MonadRWS m) => MonadRWS (ExceptT e m)
 
 instance (MonadRWS m) => MonadRWS (IdentityT m)