+++ monads-tf/Control/Monad/Reader/Class.hs
@@ -42,6 +42,7 @@ module Control.Monad.Reader.Class (
 
 import Control.Monad.Trans.Cont as Cont
 import Control.Monad.Trans.Error
+import Control.Monad.Trans.Except
 import Control.Monad.Trans.Identity
 import Control.Monad.Trans.List
 import Control.Monad.Trans.Maybe
@@ -117,6 +118,11 @@ instance (Error e, MonadReader m) => MonadReader (ErrorT e m) where
     type EnvType (ErrorT e m) = EnvType m
     ask   = lift ask
     local = mapErrorT . local
+
+instance (MonadReader m) => MonadReader (ExceptT e m) where
+    type EnvType (ExceptT e m) = EnvType m
+    ask   = lift ask
+    local = mapExceptT . local
 
 instance (MonadReader m) => MonadReader (IdentityT m) where
     type EnvType (IdentityT m) = EnvType m